Revert "Temporarily disable dmabuf video -> overlay"
Overlays aren't really overlays, but underlays. That means they're
below the primary plane, not above it, so a hole must be punched
through the primary plane. NXP did this by hacking GStreamer,
wayland-protocols and Weston. Lacking to make it all "work" is
rendering the primary plane in a format that supports alpha.

BUG: 142883165

This reverts commit e0921a74330a8c3d7dce245c70cd21f471cdde76.

Change-Id: I6cb2909842425633ef169b2144c6af37c957fb0b